% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBMinerParameters.R
\name{NBMinerParameters}
\alias{NBMinerParameters}
\title{Estimate Global Model Parameters from Data}
\usage{
NBMinerParameters(
  data,
  trim = 0.01,
  pi = 0.99,
  theta = 0.5,
  minlen = 1,
  maxlen = 5,
  rules = FALSE,
  plot = FALSE,
  verbose = FALSE,
  getdata = FALSE
)
}
\arguments{
\item{data}{the data as an object of class \link[arules:transactions]{arules::transactions}.}

\item{trim}{fraction of incidences to trim off the tail of the frequency
distribution of the data.}

\item{pi}{precision threshold \eqn{\pi}.}

\item{theta}{pruning parameter \eqn{\theta}.}

\item{minlen}{minimum number of items in found itemsets (default: 1).}

\item{maxlen}{maximum number of items in found itemsets (default: 5).}

\item{rules}{mine NB-precise rules instead of NB-frequent itemsets?}

\item{plot}{plot the model?}

\item{verbose}{use verbose output for the estimation procedure.}

\item{getdata}{also return the observed and estimated counts?}
}
\value{
An object of class \code{NBMinerParameter} for use with \code{\link[=NBMiner]{NBMiner()}}. If
\code{getdata = TRUE}, a list containing the parameter object, observed counts,
and expected counts is returned.
}
\description{
Estimate the global negative binomial data model used by NBMiner and
create an appropriate parameter object.
}
\details{
The EM algorithm estimates the global NB model because the zero class
(items that do not occur in the dataset) is not observed. The result is the
two NB parameters \eqn{k} and \eqn{a}. The value of \eqn{a} is rescaled by
dividing it by the number of incidences in the data, as required by NBMiner.
The estimated total number of items \eqn{n} is also returned.

The supplied values of \code{theta} and \code{pi} are added to the resulting
parameter object.
}
\examples{
data("Epub")

param <- NBMinerParameters(Epub, trim = 0.05, plot = TRUE, verbose = TRUE)
param

}
\references{
Michael Hahsler. A model-based frequency constraint for mining
associations from transaction data. \emph{Data Mining and Knowledge
Discovery,13(2):137-166,} September 2006.
\doi{10.1007/s10618-005-0026-2}
}
\keyword{models}
